CHET-400: Rename cature package to captor
